Making server more modular - does not build
