+++ coniks_server/src/org/coniks/coniks_server/ConiksServer.java
@@ -430,39 +430,7 @@ public class ConiksServer{
     					 CONFIG.EPOCH_INTERVAL,
     					 TimeUnit.MILLISECONDS);
 
-        ServerSocket s;
-        
-        try{
-
-            if (isFullOp) {
-                SSLServerSocketFactory sslSrvFact = 
-                (SSLServerSocketFactory)SSLServerSocketFactory.getDefault();
-                s =(SSLServerSocket)sslSrvFact.createServerSocket(CONFIG.PORT);
-            }
-            else {
-                s = new ServerSocket(CONFIG.PORT);
-
-                System.out.println("Listening for connections on port "+CONFIG.PORT+"...");
-            }            
-
-            serverLog.log("Listening for connections on port "+CONFIG.PORT+"...");
-            
-            // loop to listen for requests
-            while(true){
-                Socket c = s.accept(); // closing done by thread
-                
-                serverLog.log("Server accepted new connection.");
-                
-                ServerThread th = new ServerThread(c);
-                th.start();
-                
-            }
-        }
-        catch(Exception e){
-            serverLog.error("Exception: " + e.getMessage());
-	    e.printStackTrace();
-            System.exit(-1);
-        }
+        ServerMessaging.listen();
         
     }
     
@@ -497,577 +465,5 @@ public class ConiksServer{
         }
 
     }
-
-    /** A Thread that attempts a connection with a CONIKS client
-     *
-     * @author Marcela Melara
-     *
-     */
-    private static class ServerThread extends Thread{
-        
-        private Socket clientSocket;
-        private DataInputStream din;
-        private DataOutputStream dout;
-        private long regEpoch;
-        private int msgType;
-         
-        /** Constructor of a ServerThread
-         *
-         * @param s the client socket
-         */
-        public ServerThread(Socket c){
-            
-            if (isFullOp) {
-                this.clientSocket = (SSLSocket)c;
-            }
-            else {
-                this.clientSocket = c;
-            }
-        }
-        
-        /** Runs the ServerThread: calls the handle connection method
-         * Will have a switch statement for each message type received
-         */
-        public void run(){
-            
-            //attempt connection to the client
-            try{	
-                din = new DataInputStream(clientSocket.getInputStream());
-                dout = new DataOutputStream(clientSocket.getOutputStream());
-
-                // this will also get the message type
-                AbstractMessage clientMsg = receiveMsgProto();
-
-                if (clientMsg == null) {
-                    sendSimpleResponse(ServerUtils.RespType.MALFORMED_ERR);
-                }
-                else if (msgType == MsgType.REGISTRATION) {
-                    handleRegistrationProto((Registration) clientMsg);
-                }
-                else if (msgType == MsgType.COMMITMENT_REQ) {
-                    handleCommitmentReqProto((CommitmentReq) clientMsg);
-                }
-                else if (msgType == MsgType.KEY_LOOKUP) {
-                    handleKeyLookupProto((KeyLookup) clientMsg);
-                }
-
-                else if (msgType == MsgType.ULNCHANGE_REQ) {
-                    // TODO
-                    handleULNChangeProto((ULNChangeReq) clientMsg);
-                }
-                else if (msgType == MsgType.SIGNED_ULNCHANGE_REQ) {
-                    // TODO
-                    handleSignedULNChangeProto((SignedULNChangeReq) clientMsg);
-                }
-                
-                clientSocket.close();
-                
-            }
-            catch(IOException e){
-                if (isFullOp) {
-                    msgLog.error("Error connecting to client: "+e.getMessage());
-                }
-                else {
-                    printStatusMsg(true, "Error connecting to client: "+e.getMessage());
-                }
-                e.printStackTrace();
-            }
-
-        } //ends run()
-
-        /* Message handlers */
-        
-        /** Receives a protobuf message from the client and checks that
-         * the message is correctly formatted for the expected message type.
-         * The caller is responsible for handling the exact message type(s).
-         *
-         *@return The specific protobuf message according to the message type
-         * indicated by the client.
-         */
-        private synchronized AbstractMessage receiveMsgProto() {
-            
-            try {
-                // get the message type of the message and read in the stream
-                msgType = din.readUnsignedByte();
-                
-                if (msgType == MsgType.REGISTRATION){
-                    Registration reg = Registration.parseDelimitedFrom(din);
-                    
-                    if(!reg.hasBlob()){
-                        msgLog.log("Malformed registration message");
-                    }
-                    else {
-                        return reg;
-                    }
-                }
-                else if (msgType == MsgType.KEY_LOOKUP) {
-                    KeyLookup lookup = KeyLookup.parseDelimitedFrom(din);
-                    
-                    if(!lookup.hasName() || !lookup.hasEpoch() || 
-                       lookup.getEpoch() <= 0){
-                        msgLog.log("Malformed key lookup");
-                    }
-                    else {
-                        return lookup;
-                    }
-                }
-                else if (msgType == MsgType.COMMITMENT_REQ) {
-                    CommitmentReq commReq = CommitmentReq.parseDelimitedFrom(din);
-                    
-                    if (!commReq.hasType() || !commReq.hasEpoch() || commReq.getEpoch() <= 0) {
-                        msgLog.log("Malformed commitment request message");
-                    }
-                    else {
-                        return commReq;
-                    }
-                }
-                else if (msgType == MsgType.ULNCHANGE_REQ) {
-                    ULNChangeReq ulnChange = ULNChangeReq.parseDelimitedFrom(din);
-                    if (!ulnChange.hasName()) {
-                        msgLog.log("Malformed uln change req");
-                    }
-                    else {
-                        return ulnChange;
-                    }
-                }
-                else if (msgType == MsgType.SIGNED_ULNCHANGE_REQ) {
-                    SignedULNChangeReq sulnReq = SignedULNChangeReq.parseDelimitedFrom(din);
-                    if (!sulnReq.hasReq() || sulnReq.getSigCount() < 1 || !sulnReq.getReq().hasName()) {
-                        msgLog.log("Malformed signed uln change req");
-                    }
-                    else {
-                        return sulnReq;
-                    }
-                }
-                else {
-                    // result = ServerUtils.RespType.SERVER_ERR;
-                    msgLog.log("Unknown incoming message type");
-                }
-            }
-            catch (InvalidProtocolBufferException e) {
-                if (isFullOp) {
-                    msgLog.error("parsing a protobuf message");
-                }
-                else {
-                    printStatusMsg(true, "parsing a protobuf message");
-                }
-            }
-            catch (IOException e) {
-                 if (isFullOp) {
-                    msgLog.error("receiving data from client");
-                }
-                else {
-                    printStatusMsg(true, "receiving data from client");
-                }
-            }
-
-            // unexpected message type from the client
-            return null;
-        }
-
-        private synchronized void handleRegistrationProto(Registration reg) 
-            throws IOException{
-            msgLog.log("Handling registration message... ");
-
-            if(!reg.hasBlob()){
-                msgLog.log("Malformed registration message");
-                sendSimpleResponse(ServerUtils.RespType.MALFORMED_ERR);
-                return;
-            }
-           
-            String name =reg.getName();
-
-            // want to check first whether the name already 
-            // exists in the database before we register, if it does, reply with error
-            ServerUtils.Record r = getRecord(curEpoch);
-            RootNode root = r.getRoot();
-
-            UserLeafNode uln = getUlnFromTree(name, root);
-
-            if (uln != null) {
-                msgLog.error("Found: "+
-                             ServerUtils.bytesToHex(ServerUtils.unameToIndex(uln.getUsername()))+
-                             "\n"+uln.getUsername()+" found when trying to insert "+name);
-                sendSimpleResponse(ServerUtils.RespType.NAME_EXISTS_ERR);
-                return;
-            }
-
-            this.regEpoch = curEpoch+CONFIG.EPOCH_INTERVAL;
-                      
-            // If using a DB, insert the new user
-
-            // we register the user in the pendingQueue
-            register(name, reg.getBlob());
-
-            sendRegistrationRespResponse(regEpoch, CONFIG.EPOCH_INTERVAL);
-        
-        }
-
-        /* Helper functions for commitment requests */
-
-        // retrieves the root node and commitment signature given a specific commitment request
-        private synchronized void handleCommitmentReqProto(CommitmentReq commReq) 
-            throws IOException{
-            
-            long epoch = commReq.getEpoch();
-            // if we get a request for an epoch we haven't reached yet, return the current
-            if(epoch > curEpoch){
-                epoch = curEpoch;
-            }
-
-            msgLog.log("Getting commitment for epoch "+epoch+"...");
-
-            CommitmentReq.CommitmentType commType = commReq.getType();
-
-            // TODO: handle requests for observed commitments
-            if(commType == CommitmentReq.CommitmentType.SELF){
-                ServerUtils.Record record = getRecord(epoch);
-
-                // TODO: what to do if record not found?
-
-                RootNode root = record.getRoot();
-                byte[] str = record.getSTR();
-
-                sendCommitmentResponse(Pair.with(root, str));
-            }
-            
-        }
-
-        /* Helper functions for key lookups */
-
-        // retrieves the user leaf node given a specific key lookup
-        private synchronized void handleKeyLookupProto(KeyLookup lookup)
-            throws IOException{
-  
-            long epoch = lookup.getEpoch();
-            if(epoch > curEpoch){
-                epoch = curEpoch;
-            }
-
-            String username = lookup.getName();
- 
-            if(username.charAt(username.length()-1) == '/' ){
-                username = username.substring(0,username.length()-1);
-            }
-
-            msgLog.log("Getting key for "+username+"... ");
-
-    	    msgLog.log("SHA256 of name: " + ServerUtils.bytesToHex(ServerUtils.unameToIndex(username)));
-	    
-            ServerUtils.Record r = getRecord(epoch);
-            RootNode root = r.getRoot();	  
-
-            UserLeafNode uln = getUlnFromTree(username, root);
-
-    	    if(uln == null){
-    		msgLog.error(username + " not found...");
-                    sendSimpleResponse(ServerUtils.RespType.NAME_NOT_FOUND_ERR);
-                    return;
-    	    }
-  
-            sendAuthPathResponse(uln, root);
-        }
-
-        /* Helper functions for ULN changes (without sig) */
-
-        // retrieves the user leaf node given a specific key lookup
-        private synchronized void handleULNChangeProto(ULNChangeReq changeReq)
-            throws IOException{
-            handleULNChangeProto(changeReq, null);
-        }
-
-        // Handles a proto that might have a sig
-        private synchronized void handleULNChangeProto(ULNChangeReq changeReq, byte[] sig)
-            throws IOException{
-
-            String username = changeReq.getName();
- 
-            if(username.charAt(username.length()-1) == '/' ){
-                username = username.substring(0,username.length()-1);
-            }
-
-            ServerUtils.Record r = getRecord(curEpoch);
-            RootNode root = r.getRoot();      
-
-            // get the uln
-            UserLeafNode uln = getUlnFromTree(username, root);  
-            if(uln == null){
-                msgLog.error(username + " not found...");
-                sendSimpleResponse(ServerUtils.RespType.NAME_NOT_FOUND_ERR);
-                return;
-            }
-
-            boolean allowsUnsignedKC = changeReq.hasAllowsUnsignedKeychange() ? changeReq.getAllowsUnsignedKeychange() : uln.allowsUnsignedKeychange();
-            boolean allowsPublicLookup = changeReq.hasAllowsPublicLookup() ? changeReq.getAllowsPublicLookup() : uln.allowsPublicLookups();
-            String newBlob = changeReq.hasNewBlob() ? changeReq.getNewBlob() : uln.getPublicKey();
-            DSAPublicKey newChangeKey = changeReq.hasNewChangeKey() ? SignatureOps.makeDSAPublicKeyFromParams(changeReq.getNewChangeKey()) : uln.getChangeKey();
-            
-            ulnChange(username, newBlob, newChangeKey, allowsUnsignedKC, allowsPublicLookup, changeReq.toByteArray(), sig);
-            serverLog.log("ulnChange: " + Arrays.toString(changeReq.toByteArray()));
-            // If using a DB, insert the new user
-
-            // Send a registration response so that the client knows when to check
-            // that the changes were actually comitted
-            this.regEpoch = curEpoch+CONFIG.EPOCH_INTERVAL;
-
-            sendRegistrationRespResponse(regEpoch, CONFIG.EPOCH_INTERVAL);
-        }
-
-        /* Helper functions for ULN changes (with sig) */
-        private synchronized void handleSignedULNChangeProto(SignedULNChangeReq signedReq)
-            throws IOException{
-  
-            ULNChangeReq changeReq = signedReq.getReq();
-            byte[] reqMsg = changeReq.toByteArray();
-
-            String username = changeReq.getName();
- 
-            if(username.charAt(username.length()-1) == '/' ){
-                username = username.substring(0,username.length()-1);
-            }
-            
-            ServerUtils.Record r = getRecord(curEpoch);
-            RootNode root = r.getRoot();      
-
-            // get the uln
-            UserLeafNode uln = getUlnFromTree(username, root);  
-            if(uln == null){
-                msgLog.error(username + " not found...");
-                sendSimpleResponse(ServerUtils.RespType.NAME_NOT_FOUND_ERR);
-                return;
-            }
-
-            // verify signature
-            DSAPublicKey publicChangeKey = uln.getChangeKey();
-
-            byte[] sig = ServerUtils.intListToByteArr(new ArrayList<Integer>(signedReq.getSigList()));
-            if (!SignatureOps.verifySigFromDSA(reqMsg, sig, publicChangeKey)) {
-                msgLog.log("Failed to verify message");
-                msgLog.log("Failed sig said\n" + Arrays.toString(sig));
-                sendSimpleResponse(ServerUtils.RespType.VERIFICATION_ERR);
-                return;
-            }
-
-            // now pass to the unsigned version of the handler
-            handleULNChangeProto(changeReq, sig);
-        }
-
-        // traverses down the tree until we reach the requested user leaf node
-        // msm: this pretty much repeats the traversal in ServerOps.generateAuthPathProto
-        // so we should really find a way to remove this redundancy
-        private synchronized UserLeafNode getUlnFromTree(String username,
-                                                         RootNode root) {
-
-            // traverse based on lookup index for this name
-            byte[] lookupIndex = ServerUtils.unameToIndex(username);
-            
-            // not worth doing this recursively
-            int curOffset = 0;
-            TreeNode runner = root;
-            
-            msgLog.log("searching for: "+ServerUtils.bytesToHex(lookupIndex));
-
-            while (!(runner instanceof UserLeafNode)) {
-
-                // direction here is going to be false = left,
-                //                               true = right
-                boolean direction = ServerUtils.getNthBit(lookupIndex, curOffset);
-                
-                if (runner == null){
-                    break;
-                }
-                
-                if (runner instanceof RootNode) {
-                    
-                    RootNode curNodeR = (RootNode) runner;
-                    
-                    if(!direction){
-                        runner = curNodeR.getLeft();
-                    }
-                    else {
-                        runner = curNodeR.getRight();
-                    }
-
-                }
-
-                else {
-                    InteriorNode curNodeI = (InteriorNode) runner;
-               
-                    if(!direction){
-                        runner = curNodeI.getLeft();
-                    }                             
-                    else {
-                        runner = curNodeI.getRight();
-                    }
-
-                    // msm: rather be safe than sorry
-                    if (runner == null){
-                        break;
-                    }
-                    
-                }
-               
-                curOffset++;
-            }
-
-            // if we have found a uln, make sure it doesn't just have a common prefix
-            // with the requested node
-            if (runner != null && runner instanceof UserLeafNode) {
-                // msm: this is ugly
-                if (!username.equals(((UserLeafNode)runner).getUsername())) {
-                        return null;
-                    }
-            }
-
-            // we expect the runner to be the right uln at this point
-            return (UserLeafNode) runner;
-  
-        }
-
-        /* Message sending functions */
-
-        // send back a simple server response based on the result of the request
-        private synchronized void sendSimpleResponse(ServerUtils.RespType reqResult){
-            msgLog.log("Sending simple server response... ");
-            try{              
-                ServerResp respMsg = buildServerRespMsg(reqResult);
-                dout.writeByte(MsgType.SERVER_RESP);
-                respMsg.writeDelimitedTo(dout);
-                dout.flush();
-                din.close();
-                dout.close();        
-            }
-            catch(IOException e){
-                msgLog.error("Something went wrong while trying to send a message to the client");
-            }
-        }
-
-        // send back the commitment returned for the commitment request
-        private synchronized void sendCommitmentResponse(Pair<RootNode, 
-                                                         byte[]> commPair){
-            msgLog.log("Sending commitment response... ");
-            try{              
-                Commitment comm = buildCommitmentMsg(commPair.getValue0(), 
-                                                     commPair.getValue1());
-                byte[] rootBytes = ServerUtils.convertRootNode(commPair.getValue0());
-                msgLog.log("Root hash "+
-                             ServerUtils.bytesToHex(ServerUtils.hash(rootBytes))
-                           +"\n Epoch: "+commPair.getValue0().getEpoch()
-                           +"\n Prev: "+ServerUtils.bytesToHex(commPair.getValue0().getPrev()));
-                dout.writeByte(MsgType.COMMITMENT);
-                comm.writeDelimitedTo(dout);
-                dout.flush();
-                din.close();
-                dout.close();                
-            }
-            catch(IOException e){
-                msgLog.error("Something went wrong while trying to send a message to the client");
-            }
-        }
-
-        // send back the initial epoch and epoch interval for the newly registered user, who will cache this info
-        private synchronized void sendRegistrationRespResponse(long initEpoch, int epochInterval){
-            msgLog.log("Sending registration response... ");
-            try{              
-                RegistrationResp regResp = buildRegistrationRespMsg(initEpoch, epochInterval);
-                dout.writeByte(MsgType.REGISTRATION_RESP);
-                regResp.writeDelimitedTo(dout);
-                dout.flush();
-                din.close();
-                dout.close();                
-            }
-            catch(IOException e){
-                msgLog.error("Something went wrong while trying to send a message to the client");
-            }
-        }
-
-        // send back the authentication path based on the key lookup
-        private synchronized void sendAuthPathResponse(UserLeafNode uln, RootNode root){
-            msgLog.log("Sending authentication path response... ");
-            try{              
-                AuthPath authPath = buildAuthPathMsg(uln, root);
-                dout.writeByte(MsgType.AUTH_PATH);
-                authPath.writeDelimitedTo(dout);
-                dout.flush();
-                din.close();
-                dout.close();               
-            }
-            catch(IOException e){
-                msgLog.error("Something went wrong while trying to send a message to the client");
-            }
-        }
-
-        /* Message building functions */
-
-        // create the simple server response message
-        private ServerResp buildServerRespMsg(ServerUtils.RespType respType){
-            ServerResp.Builder respMsg = ServerResp.newBuilder();
-            switch(respType){
-            case SUCCESS:
-                respMsg.setMessage(ServerResp.Message.SUCCESS);
-                break;
-            case NAME_EXISTS_ERR:
-                respMsg.setMessage(ServerResp.Message.NAME_EXISTS_ERR);
-                break;
-            case NAME_NOT_FOUND_ERR:
-                respMsg.setMessage(ServerResp.Message.NAME_NOT_FOUND_ERR);
-                break;
-            case MALFORMED_ERR:
-                respMsg.setMessage(ServerResp.Message.MALFORMED_ERR);
-                break;
-            case COMMITMENT_RESP:
-                respMsg.setMessage(ServerResp.Message.COMMITMENT_RESP);
-                break;
-            case AUTH_PATH:
-                respMsg.setMessage(ServerResp.Message.AUTH_PATH);
-                break;
-            case VERIFICATION_ERR:
-                respMsg.setMessage(ServerResp.Message.VERIFICATION_ERR);
-                break;
-            default:
-                respMsg.setMessage(ServerResp.Message.SERVER_ERR);
-                break;                
-            }
-            return respMsg.build();
-        }
-
-        // create the commitment response message
-        private Commitment buildCommitmentMsg(RootNode root, byte[] commSig){            
-
-            Commitment.Builder commMsg = Commitment.newBuilder();
-            byte[] rootBytes = ServerUtils.convertRootNode(root);
-            byte[] rootHashBytes = ServerUtils.hash(rootBytes);
-
-            Hash.Builder rootHash = Hash.newBuilder();
-            ArrayList<Integer> rootHashList = ServerUtils.byteArrToIntList(rootHashBytes);
-           
-            if(rootHashList.size() != ServerUtils.HASH_SIZE_BYTES){
-                msgLog.error("Bad length of root hash: "+rootHashList.size());
-                return null;
-            }
-            rootHash.setLen(rootHashList.size());
-            rootHash.addAllHash(rootHashList);
-            commMsg.setEpoch(root.getEpoch());
-            ArrayList<Integer> sigList = ServerUtils.byteArrToIntList(commSig);
-            commMsg.setRootHash(rootHash.build());
-            commMsg.addAllSignature(sigList);
-            return commMsg.build();
-        }
-
-        // create the registration response message
-        private RegistrationResp buildRegistrationRespMsg(long initEpoch, int epochInterval){            
-
-            RegistrationResp.Builder regRespMsg = RegistrationResp.newBuilder();
-            regRespMsg.setInitEpoch(initEpoch);
-            regRespMsg.setEpochInterval(epochInterval);
-            return regRespMsg.build();
-        }
-
-        // create the commitment response message
-        private AuthPath buildAuthPathMsg(UserLeafNode uln, RootNode root){            
-            return ServerOps.generateAuthPathProto(uln, root);
-        }
-        
-    } //ends ServerThread class
     
 } // ends class
